+++ src/SimulationConfig/ParticleGroup.h
@@ -18,18 +18,18 @@
 class ParticleGroup {
  public:
   const int typeID; /**< Type identifier for looking up further particle properties */
-  const double spacing; /**< Describes particle density. Smaller values mean that particles are closer together. */
+  const float spacing; /**< Describes particle density. Smaller values mean that particles are closer together. */
   const Coord3D velocity; /**< Starting velocity vector of all particles of the group */
-  const double particleEpsilon; /**< Epsilon property of all particles of the group */
-  const double particleSigma; /**< Sigma property of all particles of the group */
-  const double particleMass; /**< Mass property of all particles of the group */
+  const float particleEpsilon; /**< Epsilon property of all particles of the group */
+  const float particleSigma; /**< Sigma property of all particles of the group */
+  const float particleMass; /**< Mass property of all particles of the group */
 
   ParticleGroup(int typeID,
-                double spacing,
+                float spacing,
                 const Coord3D velocity,
-                double particleEpsilon,
-                double particleSigma,
-                double particleMass)
+                float particleEpsilon,
+                float particleSigma,
+                float particleMass)
       : typeID(typeID),
         spacing(spacing),
         velocity(velocity),
